refactor: utils, util 폴더 통합
