+++ src/router/auth.ts
@@ -1,7 +1,7 @@
 import express, { Router } from "express";
 import { check, login, register } from "../controllers/auth";
 import jwtMiddleware from "../middleware/jwtMiddleware";
-import wrapAsync from "../utils/wrapAsync";
+import wrapAsync from "../util/wrapAsync";
 
 const userRouter: Router = express.Router();
