+++ src/router/record.ts
@@ -1,6 +1,6 @@
 import express, { Router } from "express";
 import { readByMonth, write } from "../controllers/record";
-import wrapAsync from "../utils/wrapAsync";
+import wrapAsync from "../util/wrapAsync";
 
 const recordRouter: Router = express.Router();
